+++ com.unity.render-pipelines.high-definition/Runtime/Lighting/Light/HDAdditionalLightData.cs
@@ -195,11 +195,11 @@ namespace UnityEngine.Rendering.HighDefinition
         /// </summary>
         public float innerSpotPercent01 => innerSpotPercent / 100f;
 
-        [Range(0.0f, 1.0f)]
+        [Range(0.0f, 16.0f)]
         [SerializeField, FormerlySerializedAs("lightDimmer")]
         float m_LightDimmer = 1.0f;
         /// <summary>
-        /// Get/Set the light dimmer.
+        /// Get/Set the light dimmer / multiplier, between 0 and 16. 
         /// </summary>
         public float lightDimmer
         {
@@ -209,14 +209,14 @@ namespace UnityEngine.Rendering.HighDefinition
                 if (m_LightDimmer == value)
                     return;
 
-                m_LightDimmer = Mathf.Clamp01(value);
+                m_LightDimmer = Mathf.Clamp(value, 0.0f, 16.0f);
             }
         }
 
-        [Range(0.0f, 1.0f), SerializeField, FormerlySerializedAs("volumetricDimmer")]
+        [Range(0.0f, 16.0f), SerializeField, FormerlySerializedAs("volumetricDimmer")]
         float m_VolumetricDimmer = 1.0f;
         /// <summary>
-        /// Get/Set the light dimmer on volumetric effects, between 0 and 1.
+        /// Get/Set the light dimmer / multiplier on volumetric effects, between 0 and 16.
         /// </summary>
         public float volumetricDimmer
         {
@@ -226,7 +226,7 @@ namespace UnityEngine.Rendering.HighDefinition
                 if (m_VolumetricDimmer == value)
                     return;
 
-                m_VolumetricDimmer = Mathf.Clamp01(value);
+                m_VolumetricDimmer = Mathf.Clamp(value, 0.0f, 16.0f);
             }
         }